Add some more monster flags
Add flags to remove the amount of monster names hard wired in
is_whatever macros. This requires extending mflags3 to a long.

+++ include/permonst.h
@@ -55,11 +55,11 @@ struct permonst {
         cnutrit;                /* its nutritional value */
     uchar msound;               /* noise it makes (6 bits) */
     uchar msize;                /* physical size (3 bits) */
-    unsigned long mresists;    /* resistances */
+    unsigned long mresists;     /* resistances */
     uchar mconveys;             /* conveyed by eating */
     unsigned long mflags1,      /* boolean bitflags */
         mflags2;                /* more boolean bitflags */
-    unsigned short mflags3;     /* yet more boolean bitflags */
+    unsigned long mflags3;      /* yet more boolean bitflags */
 #ifdef TEXTCOLOR
     uchar mcolor; /* color to use */
 #endif
